+++ tests/conftest.py
@@ -48,6 +48,7 @@ class FixturesSettingsTestMixin(object):
 
 @pytest.fixture(scope='session')
 def temp_builds_dir(tmpdir_factory):
+    """Prepare a temporary build directory"""
     fn = tmpdir_factory.mktemp('builds')
     return fn