+++ tests/test_500_cli_002_watch.py
@@ -1,0 +1,6 @@
+# -*- coding: utf-8 -*-
+"""
+Watch commandline is currently not tested because of some limits with
+``CliRunner`` that is not able to communicate with invoked command, so we
+can't send interrupt with "CTRL+C". Also it may low down tests running.
+"""